AI_services/app/services/burn_out_service/api/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Pydantic for validation
pydantic==2.5.3
pydantic[email]==2.5.3

# CORS middleware
python-dotenv==1.0.0

# Already have these from main system
# sqlalchemy
# psycopg[binary]
# langchain-ollama
# langchain-postgres
